refactor: remove web secret encryption leftovers
